Tons of changes, mostly using Gibbs
still in a rough state. Most relevant changes
perform max_attempts steps in a row (doesn't exit as soon as it
succeeds)
uses Gibbs distr. to pick the sample to throw away (the criterion is
still that the new energy shouldn't be worse than E_worst)
also (optionally) use Gibbs to compute the barycenter (doesn't seem to
help much actually)
shrinks the distances and lowers the temp. gradually (every 1000 steps)